Moved get and setChildren to parent Node class.
