+++ src/core/QOANode.java
@@ -1,8 +1,5 @@
 package core;
 
-import java.util.ArrayList;
-import java.util.List;
-
 import com.google.gson.annotations.Expose;
 
 /**QuantityOfArgumentsNode*/
@@ -10,14 +7,11 @@ class QOANode extends Node {
   @Expose
   private int qoa;
 
-  @Expose
-  private List<PositionNode> children;
-
   public QOANode(int quantity) {
     this.qoa = quantity;
-    this.children = new ArrayList<PositionNode>();
+    this.children = new Node[quantity];
     for (int i = 0; i < quantity; ++i) {
-      this.children.add(new PositionNode(i + 1));
+      this.children[i] = new PositionNode(i + 1);
     }
   }
 
@@ -27,21 +21,10 @@ class QOANode extends Node {
     
     increment();
     for (int i = 0; i < children.length; ++i) {
-      PositionNode position = this.children.get(i);
+      PositionNode position = (PositionNode) this.children[i];
       FormulaToken child = children[i];
       position.add(child, example); 
     }
-  }
-  
-  public void setChildren() {
-    for (PositionNode child : children) {
-      child.setChildren();
-    }
-  }
-
-  @Override
-  public Node[] getChildren() {
-    return (Node[]) children.toArray();
   }
 
   @Override
